// packages/labextension/src/model/assignmentSettings.ts
/**
 * Grader Extension API Schemas
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SubmissionPeriod } from './submissionPeriod';

export interface AssignmentSettings {
  deadline?: string;
  max_submissions?: number;
  allowed_files?: Array<string>;
  late_submission?: Array<SubmissionPeriod>;
  autograde_type?: AssignmentSettings.AutogradeTypeEnum;
  group?: string;
  cell_timeout?: number;
}
export namespace AssignmentSettings {
  export const AutogradeTypeEnum = {
    Auto: 'auto',
    FullAuto: 'full_auto',
    Unassisted: 'unassisted'
  } as const;
  export type AutogradeTypeEnum =
    (typeof AutogradeTypeEnum)[keyof typeof AutogradeTypeEnum];
}